Removed BackenApplicationTests.java test suite
Due to the way that Gradle runs tests if there is a test suite that
references a set of tests those tests will be run once in the suite
and once on their own. This caused some tests to fail when they were
run twice in a row as they depended on the accounts they were creating
not existing yet.